+++ src/Stormancer.Plugins/Party/Stormancer.Server.Plugins.Party/Changelog.rst
@@ -16,9 +16,11 @@ Changed
 - Remove Dependency to Stormancer.Abstractions.Server because it's implicit.
 - Send IndexedDocument as part of party updated events.
 - Add configuration and event handler extensibility points to control when ready state is reset, validate ready state changes and user data updates.
+- Parties are now destroyed immediatly after the last player left to prevent players from joining afterwards.
 Added
 *****
 - Add arguments to set member data on party creation or on joining.
+
 
 4.3.1.13
 --------